Attach go debugger when it is ready instead of aon a timeout
